Move CA2213: DisposableFieldsShouldBeDisposed to the new location
